uint64_t type changed

+++ network/common.cpp
@@ -34,9 +34,9 @@ void time(const char * timeStr, uint64_t delta_time){
 	fprintf(stderr, "[%s] %s %llu us\n", CLOCK, timeStr, delta_time);
 }
 
-void cpu(const char * cpuStr, double cpuTime){
+void cpu(const char * cpuStr, uint64_t cpuTime){
 	if(DEBUG)
-	fprintf(stderr, "[%s] %s %f us\n", CPU_USE, cpuStr, cpuTime);
+	fprintf(stderr, "[%s] %s %llu us\n", CPU_USE, cpuStr, cpuTime);
 }
 
 void cipher(const char * cpuStr, const char* cipher){
